copy: mv space definitions to space_forward.h
Eventually, hopefully the current space.h can go away, and this can
be renamed to space.h
